reorganized files, added form container and add button container
