+++ src/javascripts/components/domBuilder.js
@@ -3,6 +3,8 @@ const domBuilder = () => {
   <div id="navigation"></div>
   <div id="main-container">
     <div id="title-container"></div>
+    <div id="add-button"></div>
+    <div id="form-container"></div>
     <div id="card-container"></div>
   </div>`;
 };
