+++ src/javascripts/helpers/events/navEvents.js
@@ -1,8 +1,8 @@
 import { emptyEquipment, showEquipment } from '../../components/equipment';
-import getEquipment from '../equipmentData';
+import getEquipment from '../data/equipmentData';
 import showVendors from '../../components/vendor';
 import { showRides, emptyRides } from '../../components/rides';
-import getRides from '../ridesData';
+import getRides from '../data/ridesData';
 import { emptyDinos, showDinos } from '../../components/cards/dinos';
 import getDinos from '../data/dinoData';
 import { showStaff, emptyStaff } from '../../components/cards/staff';
